+++ src/codec/tests.rs
@@ -9,7 +9,7 @@ fn decoder_simple() {
     let buf = &mut BytesMut::new();
     buf.reserve(1024);
 
-    buf.put("ping\r\npOnG\r\n");
+    buf.put("ping\r\npOnG\r\n".as_bytes());
     assert_eq!(
         codec.decode(buf).unwrap().unwrap().unwrap(),
         ServerMessage::Ping
@@ -20,11 +20,14 @@ fn decoder_simple() {
     );
     assert!(codec.decode(buf).unwrap().is_none());
 
-    buf.put("+Ok\r\nbad message here\r\n");
-    buf.put(&format!(
-        "-err '{} test.x.*.y.>'\r\n",
-        PERMISSIONS_VIOLATION_FOR_SUBSCRIPTION
-    ));
+    buf.put("+Ok\r\nbad message here\r\n".as_bytes());
+    buf.put(
+        format!(
+            "-err '{} test.x.*.y.>'\r\n",
+            PERMISSIONS_VIOLATION_FOR_SUBSCRIPTION
+        )
+        .as_bytes(),
+    );
     assert_eq!(
         codec.decode(buf).unwrap().unwrap().unwrap(),
         ServerMessage::Ok
@@ -37,14 +40,14 @@ fn decoder_simple() {
     );
 
     // Invalid utf8
-    buf.put(vec![0, 159, 146, 150, 10]);
+    buf.put(vec![0, 159u8, 146, 150, 10].as_slice());
     assert!(codec.decode(buf).unwrap().unwrap().is_err());
 
-    buf.put("pi");
+    buf.put("pi".as_bytes());
     assert!(codec.decode(buf).unwrap().is_none());
-    buf.put("ng\r");
+    buf.put("ng\r".as_bytes());
     assert!(codec.decode(buf).unwrap().is_none());
-    buf.put("\n");
+    buf.put("\n".as_bytes());
     assert_eq!(
         codec.decode(buf).unwrap().unwrap().unwrap(),
         ServerMessage::Ping
@@ -60,7 +63,8 @@ fn decoder_info() {
     buf.put(
         "INFO {\"server_id\":\"Zk0GQ3JBSrg3oyxCRRlE09\",\"version\":\"1.2.0\",\"proto\":1,\"\
          go\":\"go1.10.3\",\"host\":\"0.0.0.0\",\"port\":4222,\"max_payload\":1048576,\"\
-         client_id\":2392}\r\n",
+         client_id\":2392}\r\n"
+            .as_bytes(),
     );
     assert_eq!(
         codec.decode(buf).unwrap().unwrap().unwrap(),
@@ -87,15 +91,15 @@ fn decoder_msg() {
     let buf = &mut BytesMut::new();
     buf.reserve(1024);
 
-    buf.put("ping\r\nmsg test 0 12\r\nhello w");
+    buf.put("ping\r\nmsg test 0 12\r\nhello w".as_bytes());
     assert_eq!(
         codec.decode(buf).unwrap().unwrap().unwrap(),
         ServerMessage::Ping
     );
     assert!(codec.decode(buf).unwrap().is_none());
-    buf.put("orld!\r\n");
-    buf.put("msg test 0 5\r\nshort\r\n");
-    buf.put("msg test 0 0\r\n\r\n");
+    buf.put("orld!\r\n".as_bytes());
+    buf.put("msg test 0 5\r\nshort\r\n".as_bytes());
+    buf.put("msg test 0 0\r\n\r\n".as_bytes());
     assert_eq!(
         codec.decode(buf).unwrap().unwrap().unwrap(),
         ServerMessage::Msg(Msg::new(
@@ -124,7 +128,7 @@ fn decoder_msg() {
         ))
     );
 
-    buf.put("msg test 0 4\r\nhello world\r\nping\r\n");
+    buf.put("msg test 0 4\r\nhello world\r\nping\r\n".as_bytes());
     assert!(codec.decode(buf).unwrap().unwrap().is_err());
     assert!(codec.decode(buf).unwrap().unwrap().is_err());
     assert_eq!(
@@ -132,7 +136,7 @@ fn decoder_msg() {
         ServerMessage::Ping
     );
 
-    buf.put("msg test 0 reply 13\r\nhello\r\nworld!\r\n");
+    buf.put("msg test 0 reply 13\r\nhello\r\nworld!\r\n".as_bytes());
     assert_eq!(
         codec.decode(buf).unwrap().unwrap().unwrap(),
         ServerMessage::Msg(Msg::new(