+++ tests/request.rs
@@ -1,5 +1,6 @@
 mod common;
 
+use common::NatsServer;
 use futures::stream::StreamExt;
 use rants::{Client, Subject};
 
@@ -18,7 +19,11 @@ async fn make_subscription(client: Client, subject: &Subject) {
     });
 }
 
-async fn main() {
+#[tokio::test(threaded_scheduler)]
+async fn request() {
+    common::init();
+    let _nats_server = NatsServer::new(&[]).await;
+
     let address = "127.0.0.1".parse().unwrap();
     let client = Client::new(vec![address]);
     client.connect_mut().await.echo(true);
@@ -41,9 +46,4 @@ async fn main() {
     assert_eq!(&reply, "the reply");
 
     client.disconnect().await;
-}
-
-#[test]
-fn request() {
-    common::run_integration_test(main(), &[]);
 }