+++ tests/wild_card_subject.rs
@@ -1,9 +1,14 @@
 mod common;
 
+use common::NatsServer;
 use futures::stream::StreamExt;
 use rants::Client;
 
-async fn main() {
+#[tokio::test(threaded_scheduler)]
+async fn wild_card_subject() {
+    common::init();
+    let _nats_server = NatsServer::new(&[]).await;
+
     let address = "127.0.0.1".parse().unwrap();
     let client = Client::new(vec![address]);
     client.connect_mut().await.echo(true);
@@ -35,9 +40,4 @@ async fn main() {
 
     assert!(wild_card_subscription.next().await.is_none());
     assert!(a_subscription.next().await.is_none());
-}
-
-#[test]
-fn wild_card_subject() {
-    common::run_integration_test(main(), &[]);
 }